pyproject.toml
[build-system]
requires = [
  "setuptools>=77",
  "setuptools_scm[toml]>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "esslivedata"
description = "Live data reduction visualisation framework for ESS."
authors = [{ name = "Scipp contributors" }]
license = "BSD-3-Clause"
license-files = ["LICENSE"]
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
requires-python = ">=3.11"

# IMPORTANT:
# Run 'tox -e deps' after making changes here. This will update requirement files.
# Make sure to list one dependency per line.
dependencies = [
  "PyYAML",
  "confluent_kafka",
  "ess-streaming-data-types",
  "essreduce>=25.07.0",
  "jinja2",
  "pydantic>=2",
  "scipp>=25.02.0",  # Improved performance of scipp.group
  "scippneutron>=25.07.0",  # For DREAM unit conversion fixes
  "pooch",
]

dynamic = ["version"]

[project.optional-dependencies]
http = [
    "fastapi",
    "requests",
    "uvicorn",
]
test = [
    "essdiffraction>=25.08.0",
    "esssans",
    "essspectroscopy>=25.10.0",
    "holoviews",
    "markdown-it-py==3.0.0",  # Work around myst-parser upper bound
    "panel",
    "param",
    "pytest",
    "pytest-benchmark",
    "scippneutron>=25.07.0",
]
bifrost = [
    "essspectroscopy>=25.10.0",
]
dream = [
    "essdiffraction>=25.08.0",
]
loki = [
    "esssans",
]
odin = [
    "essimaging",
]
dashboard = [
  "bokeh",
  "gunicorn",
  "holoviews",
  "panel",
  "param",
]

[project.scripts]
ess-livedata-make-geometry-nexus = "ess.livedata.scripts.make_geometry_nexus:main"

[project.urls]
"Bug Tracker" = "https://github.com/scipp/esslivedata/issues"
"Documentation" = "https://scipp.github.io/esslivedata"
"Source" = "https://github.com/scipp/esslivedata"

[tool.setuptools_scm]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = """
--strict-config
--strict-markers
--import-mode=importlib
--benchmark-skip
-ra
-v
"""
testpaths = "tests"
filterwarnings = [
  "error",
]
pythonpath = "tests"

[tool.ruff]
line-length = 88
extend-include = ["*.ipynb"]
extend-exclude = [
    ".*", "__pycache__", "build", "dist", "install",
]

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules/
select = ["B", "C4", "DTZ", "E", "F", "G", "I", "PERF", "PGH", "PT", "PYI", "RUF", "S", "T20", "UP", "W"]
ignore = [
    # Conflict with ruff format, see
    # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "COM812", "COM819", "D206", "D300", "E111", "E114", "E117", "ISC001", "ISC002", "Q000", "Q001", "Q002", "Q003", "W191",
]
fixable = ["B010", "I001", "PT001", "RUF022"]
isort.known-first-party = ["ess.livedata"]
pydocstyle.convention = "numpy"

[tool.ruff.lint.per-file-ignores]
# those files have an increased risk of relying on import order
"tests/*" = [
    "S101",  # asserts are fine in tests
    "B018",  # 'useless expressions' are ok because some tests just check for exceptions
]
"*.ipynb" = [
    "E501",  # longer lines are sometimes more readable
    "F403",  # *-imports used with domain types
    "F405",  # linter may fail to find names because of *-imports
    "I",  # we don't collect imports at the top
    "S101",  # asserts are used for demonstration and are safe in notebooks
    "T201",  # printing is ok for demonstration purposes
]

[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]
warn_unreachable = true

[tool.setuptools.package-data]
"ess.livedata.config.defaults" = ["*.yaml", "*.yaml.jinja"]

[tool.codespell]
ignore-words-list = [
    # Codespell wants "socioeconomic" which seems to be the standard spelling.
    # But we use the word in our code of conduct which is the contributor covenant.
    # Let's not modify it if we don't have to.
    "socio-economic",
]

[tool.pylint.main]
load-plugins = ["pylint_pydantic"]
